Used named imports for react-router in react-router-config (#6291)
I was recently doing some bundle size analysis using source-map-explorer
and noticed that we had a little bit of duplication of react-router
code. It seems as though our import of react-router-config was bringing
in modules from react-router/Foo, while our named imports from
react-router was bringing in modules from react-router/es/Foo.

As a workaround, I've set up a webpack alias to avoid the duplication,
but I think this should be solved in this package so that everyone can
benefit.

I believe this is similar to:

- https://github.com/ReactTraining/react-router/pull/5589